Move `Diagnostics` class into `diagnostics/__init__.py`
